+++ packages/react/src/components/Listbox/ListboxOption.tsx
@@ -14,6 +14,7 @@ interface ListboxOptionProps
   extends PolymorphicProps<React.HTMLAttributes<HTMLElement>> {
   value?: ListboxValue;
   disabled?: boolean;
+  selected?: boolean;
   activeClass?: string;
 }
 
@@ -30,6 +31,7 @@ const ListboxOption = forwardRef<HTMLElement, ListboxOptionProps>(
       children,
       value,
       disabled,
+      selected: selectedProp,
       activeClass = 'ListboxOption--active',
       onClick,
       ...props
@@ -39,10 +41,14 @@ const ListboxOption = forwardRef<HTMLElement, ListboxOptionProps>(
     const { active, selected, setOptions, onSelect } = useListboxContext();
     const listboxOptionRef = useSharedRef<HTMLElement>(ref);
     const [id] = propId ? [propId] : useId(1, 'listbox-option');
-    const isActive =
-      active !== null && active.element === listboxOptionRef.current;
+    const isActive = active?.element === listboxOptionRef.current;
     const isSelected =
-      selected !== null && selected.element === listboxOptionRef.current;
+      typeof selectedProp === 'boolean'
+        ? selectedProp
+        : selected !== null &&
+          !!selected.find(
+            (option) => option.element === listboxOptionRef.current
+          );
     const optionValue =
       typeof value !== 'undefined'
         ? value
@@ -98,7 +104,7 @@ const ListboxOption = forwardRef<HTMLElement, ListboxOptionProps>(
         onSelect({ element: listboxOptionRef.current, value: optionValue });
         onClick?.(event);
       },
-      [optionValue]
+      [optionValue, onSelect, onClick, disabled]
     );
 
     return (