+++ packages/react/src/components/MenuItem/index.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import sinon from 'sinon';
-import { screen, render } from '@testing-library/react';
+import { screen, render, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import MenuItem from '../MenuItem';
 import { axe } from 'jest-axe';
@@ -9,9 +9,11 @@ const user = userEvent.setup();
 
 test('clicks first direct child link given a click', async () => {
   const onClick = sinon.spy();
+  // Note: Using a hash link instead of a url link because jsdom doesn't correctly
+  // support navigation and throws a noisy console error we don't care about
   render(
     <MenuItem>
-      <a href="/foo" onClick={onClick}>
+      <a href="#foo" onClick={onClick}>
         Foo
       </a>
     </MenuItem>
@@ -29,6 +31,7 @@ test('calls onClick prop', async () => {
   await user.click(screen.getByText('BOOGNISH'));
   expect(onClick.calledOnce).toBeTruthy();
 });
+
 test('clicks the menuitem given enter/space keydowns', async () => {
   const onClick = sinon.spy();
   const handleKeyDown = (event: React.KeyboardEvent) => {
@@ -67,6 +70,9 @@ test('should return no axe violations', async () => {
       <MenuItem>Foo</MenuItem>
     </ul>
   );
+  await waitFor(() => {
+    expect(container).toBeInTheDocument();
+  });
   const results = await axe(container);
   expect(results).toHaveNoViolations();
 });